% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcleaner.R
\docType{package}
\name{dcleaner}
\alias{dcleaner}
\title{dcleaner: A package for cleanning and mungling data, importing and exporting data to DB.}
\description{
The dcleaner package provides three categories of important functions:
load_table, import_table, clean_table
}
\section{Loading files}{

dcleaner could work with excel file with additional files which describes their schemas.
}

